[feat] 2차 Refactoring 진행 중
