--- src/main/java/camp/nextstep/edu/nextstep8/ladder/Ladder.java
@@ -1,11 +1,0 @@
-package camp.nextstep.edu.nextstep8.ladder;
-
-public class Ladder {
-    private final Point[][] map;
-    public Ladder(Point[][] map) {
-        this.map = map;
-    }
-    public Point[][] getMap() {
-        return this.map;
-    }
-}
