+++ src/main/java/camp/nextstep/edu/nextstep8/ladder/entity/Ladder.java
@@ -1,0 +1,16 @@
+package camp.nextstep.edu.nextstep8.ladder.entity;
+
+import java.util.Collections;
+import java.util.List;
+
+public class Ladder {
+    private List<Row> rows;
+
+    public Ladder(List<Row> rows) {
+        this.rows = rows;
+    }
+
+    public List<Row> getRows() {
+        return Collections.unmodifiableList(rows);
+    }
+}
