+++ src/main/java/camp/nextstep/edu/nextstep8/ladder/entity/Row.java
@@ -1,0 +1,8 @@
+package camp.nextstep.edu.nextstep8.ladder.entity;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class Row {
+    private List<Point> points = new ArrayList<>();
+}
